apps/api/requirements.txt
# Core FastAPI & async
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.40.0

# Database & ORM
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
pgvector>=0.3.0
alembic>=1.13.0                 # for migrations (alternative to Prisma)

# Auth & Security
argon2-cffi>=23.1.0
pyjwt>=2.9.0
pyotp>=2.9.0
cryptography>=43.0.0
passlib[bcrypt]>=1.7.4

# AI & Agents (LangChain + xAI)
langchain-core>=0.2.0
langchain-xai>=0.1.0            # official xAI integration
langgraph>=0.1.0
tenacity>=8.5.0

# Stripe & Billing
stripe>=11.0.0

# Email
sendgrid>=6.11.0

# Tasks & Queue
celery[redis]>=5.4.0
redis>=5.0.0

# Rate limiting & security
slowapi>=0.1.9

# Observability
sentry-sdk[fastapi]>=2.0.0

# Utils
pydantic-settings>=2.5.0
httpx>=0.27.0
qrcode[pil]>=8.0.0
python-multipart>=0.0.9

# Testing & dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0